link/requirements.txt
Django==1.8.9
Markdown==2.6.5
chardet==2.3.0
django-filter==0.12.0
djangorestframework==3.3.2
psycopg2==2.6.1
